Merge master with kl0tl/es-modules, then run `nix-shell --command make`
When merging, I checked most packages to see whether the repo and/or 
version a package uses in the es-modules branch is significantly 
different from the current master. If it was, I defaulted to the 
non-0.14.0-compatible master branch rather than the es-modules' branch's 
change. This will require additional work on our part before this 
package set compiles. However, it means bugs and other important changes 
won't be removed from those libraries.