Update CI: test if packages updated to 0.14 rc build without err/warn (#716)
* Update CI: test if packages updated to 0.14 rc build without err/warn

* Override the metadata.version so package set works with spago

* Install psa so we can use `--strict` to fail on compiler warnings

* Verify CI fails when warnings are emitted by using older prelude

The older version of prelude uses a commit where all compiler errors are fixed but not all compiler warnings.

* Install packages via `updatedLibs.dhall`; test for errors & warnings via `usePsa.dhall`

CI will fail when compilation errors are emitted, but it won't fail when compiler warnings are emitted.
Due to the breaking changes, there will be many compiler warnings. Thus, we want CI to also fail when warnings are emitted. 

We can use `purescript-psa` to fail the build via the `--strict` flag. However, this only works on warnings emitted by
files in the `src` directory, not files found in dependency folders (e.g.  the `.spago` directory).

We get around this issue doing the following three steps:
1. Use spago via its `updatedLibs.dhall` file to download all the packages we need to test
2. Rename the `.spago` folder to `src`
3. Use spago via its `usePsa.dhall` file to compile all the packages that are now in the `src` directory
4. Spago calls `psa` behind the scenes with the `--strict` flag
5. If there are 0 warnings and errors, CI will pass. Otherwise, CI will fail.

* Refactor spago.dhall configuration

Due to the above workaround, we need to use two separate `spago.dhall` configuration files.
If we forget to override the packages' `metadata.version` field, spago will fail because of its internal check
and thus CI will also fail. Rather than needing to remember to update this in two separate places (i.e. updatedLibs.dhall and 
usePsa.dhall), we update it once shared.dhall.

* Update `prelude` back to `master` so CI passes again

Now that we've verified that CI fails when compiler warnings are emitted, we can revert our previous commit
and use the version of prelude that emits 0 compiler warnings and errors.